Dirty fix to work with plasticboy/vim-markdown
This issue was originally mentioned in
https://github.com/vim-pandoc/vim-criticmarkup/issues/11 and later in
https://github.com/plasticboy/vim-markdown/issues/481
Albeit being quite a manual and dirty fix, this commit fixes the
problems for common syntax regions.
If others are found to conflict they may simply also be appended to the
`containedin` argument list.

TODO: Is it possible to update a defined syntax region element's
attribute after its original definition? If yes, the appending of these
syntax groups could be outsourced to a later point or even a separate
configuration file. However, I haven't found a way of achieving this.